we-mqtt-qa/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
dnspython==1.16.0
eventlet==0.28.0
greenlet==0.4.17
idna==2.10
monotonic==1.5
mypy-extensions==0.4.3
paho-mqtt==1.5.1
qwikidata==0.4.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sparql-client==3.8
urllib3==1.25.10